Refactor chiral conversion code into chiral_utils.py
