+++ timemachine/fe/single_topology.py
@@ -1,23 +1,26 @@
 import warnings
-from collections.abc import Iterable
 from enum import IntEnum
 from functools import partial
 from typing import Callable, Collection, Dict, FrozenSet, List, Optional, Tuple, TypeVar, Union, cast
 
 import jax
 import jax.numpy as jnp
-import networkx as nx
 import numpy as np
 from numpy.typing import NDArray
 from rdkit import Chem
 
 from timemachine.constants import DEFAULT_CHIRAL_ATOM_RESTRAINT_K, DEFAULT_CHIRAL_BOND_RESTRAINT_K
-from timemachine.fe import chiral_utils, interpolate, model_utils, topology, utils
-from timemachine.fe.chiral_utils import ChiralRestrIdxSet
+from timemachine.fe import interpolate, model_utils, topology, utils
+from timemachine.fe.chiral_utils import (
+    assert_chiral_consistency_and_validity,
+    setup_dummy_bond_and_chiral_interactions,
+    setup_end_state_harmonic_bond_and_chiral_potentials,
+)
 from timemachine.fe.dummy import canonicalize_bond, generate_anchored_dummy_group_assignments
 from timemachine.fe.lambda_schedule import construct_pre_optimized_relative_lambda_schedule
 from timemachine.fe.system import HostGuestSystem, VacuumSystem
 from timemachine.fe.topology import get_ligand_ixn_pots_params
+from timemachine.fe.utils import recursive_map
 from timemachine.potentials import (
     BoundPotential,
     ChiralAtomRestraint,
@@ -41,26 +44,6 @@ class MissingAngleError(RuntimeError):
 
 class ChargePertubationError(RuntimeError):
     pass
-
-
-class ChiralConversionError(RuntimeError):
-    pass
-
-
-def recursive_map(items, mapping):
-    """recursively replace items in a list of tuple
-    mapping = np.arange(100)[::-1]
-    items = [[0,2,3], [5,1,[2,5,6]], 3]
-    result = recursive_map(items, mapping)
-    # ((99, 97, 96), (94, 98, (97, 94, 93)), 96)
-    """
-    if isinstance(items, Iterable):
-        res = []
-        for item in items:
-            res.append(recursive_map(item, mapping))
-        return tuple(res)
-    else:
-        return mapping[items]
 
 
 def setup_dummy_interactions_from_ff(
@@ -95,87 +78,6 @@ def setup_dummy_interactions_from_ff(
         nbr_anchor_atom,
         core_atoms,
     )
-
-
-def setup_dummy_bond_and_chiral_interactions_from_ff(
-    ff, mol, dummy_group, root_anchor_atom, core_atoms, chiral_atom_k, chiral_bond_k
-):
-    """
-    Setup interactions involving atoms in a given dummy group.
-    """
-    top = topology.BaseTopology(mol, ff)
-
-    bond_params, hb = top.parameterize_harmonic_bond(ff.hb_handle.params)
-    chiral_atom_potential, _ = top.setup_chiral_restraints(chiral_atom_k, chiral_bond_k)
-    chiral_atom_idxs = chiral_atom_potential.potential.idxs
-    chiral_atom_params = chiral_atom_potential.params
-
-    return setup_dummy_bond_and_chiral_interactions(
-        hb.idxs,
-        bond_params,
-        chiral_atom_idxs,
-        chiral_atom_params,
-        dummy_group,
-        root_anchor_atom,
-        core_atoms,
-    )
-
-
-def setup_dummy_bond_and_chiral_interactions(
-    bond_idxs,
-    bond_params,
-    chiral_atom_idxs,
-    chiral_atom_params,
-    dummy_group,
-    root_anchor_atom,
-    core_atoms,
-):
-    assert root_anchor_atom in core_atoms
-
-    dummy_bond_idxs = []
-    dummy_bond_params = []
-    dummy_chiral_atom_idxs = []
-    dummy_chiral_atom_params = []
-
-    # dummy_group may be a set in certain cases, so sanity check.
-
-    assert len(dummy_group) == len(list(dummy_group))
-    dummy_group = list(dummy_group)
-
-    # dummy group and anchor
-    dga = dummy_group + [root_anchor_atom]
-
-    # copy interactions that involve only root_anchor_atom
-    for idxs, params in zip(bond_idxs, bond_params):
-        if all([a in dga for a in idxs]):
-            dummy_bond_idxs.append(tuple([int(x) for x in idxs]))  # tuples are hashable etc.
-            dummy_bond_params.append(params)
-
-    # certain configuration of chiral states are symmetrizable
-    # . means a bond involving at least 1 dummy atom
-    # | means a bond involving only core atoms
-    #
-    #        all allowed geometries              |  disallowed geometry
-    #     center is core     center is not core  |
-    #    d      c      c      d      d      d    |    c      c
-    #    .      |      |      .      .      .    |    .      |
-    #    c      c      c      d      d      d    |    d      c
-    #   . .    . .    / .    . .    . .    . .   |   . .    / \
-    #  d   d  d   d  c   d  c   d  c   c  d   d  |  c   c  c   c (only core)
-    dgc = dummy_group + list(core_atoms)
-    for idxs, params in zip(chiral_atom_idxs, chiral_atom_params):
-        center, i, j, k = idxs
-        if all([a in dgc for a in idxs]):
-            # non center dummy atom count
-            ncda_count = sum([a in dummy_group for a in (i, j, k)])
-            if ncda_count == 1 or ncda_count == 2 or ncda_count == 3:
-                assert not all(a in core_atoms for a in idxs)
-                dummy_chiral_atom_idxs.append(tuple(int(x) for x in idxs))
-                dummy_chiral_atom_params.append(params)
-
-    bonded_idxs = (dummy_bond_idxs, dummy_chiral_atom_idxs)
-    bonded_params = (dummy_bond_params, dummy_chiral_atom_params)
-    return bonded_idxs, bonded_params
 
 
 def setup_dummy_interactions(
@@ -357,157 +259,6 @@ def canonicalize_improper_idxs(idxs) -> Tuple[int, int, int, int]:
             break
 
     return (i, *cw_items[idx])
-
-
-def get_num_connected_components(num_atoms: int, bonds: Collection[Tuple[int, int]]) -> int:
-    g = nx.Graph()
-    g.add_nodes_from(range(num_atoms))
-    g.add_edges_from(bonds)
-    return len(list(nx.connected_components(g)))
-
-
-def canonicalize_chiral_atom_idxs(idxs):
-    i, j, k, l = idxs
-    rotations = [(j, k, l), (l, j, k), (k, l, j)]
-    jj, kk, ll = min(rotations)
-    return i, jj, kk, ll
-
-
-def setup_end_state_harmonic_bond_and_chiral_potentials(
-    ff, mol_a, mol_b, core, a_to_c, b_to_c
-) -> Tuple[BoundPotential[HarmonicBond], BoundPotential[ChiralAtomRestraint], BoundPotential[ChiralBondRestraint]]:
-    """
-    Setup end-state potentials to verify chiral correctness for mol_a with dummy atoms of mol_b attached. The mapped indices will correspond
-    to the alchemical molecule with dummy atoms. Note that the bond, chiral atom and chiral bond idxs are canonicalized.
-
-    This code is identical to setup_end_state, but only handles chiral potentials to be used to quickly verify
-    core mappings in verify_chiral_consistency_of_core
-
-    Parameters
-    ----------
-    ff: forcefield.Forcefield
-        Forcefield used to parameterize the molecule
-
-    mol_a: Chem.Mol
-        Fully interacting molecule
-
-    mol_b: Chem.Mol
-        Molecule providing the dummy atoms.
-
-    core: list of 2-tuples
-        Each pair is an atom mapping from mol_a into mol_b
-
-    a_to_c: dict or array, supports []
-        mapping from a into a common core idx
-
-    b_to_c: dict or array, supports []
-        mapping from b into a common core idx
-
-    Returns
-    -------
-    Tuple of bound HarmonicBond, ChiralAtomRestraint and ChiralBondRestraint
-        The potentials involved in checking the correctness and validity of chirality at endstates.
-
-    """
-    all_dummy_bond_idxs, all_dummy_bond_params = [], []
-    all_dummy_chiral_atom_idxs, all_dummy_chiral_atom_params = [], []
-
-    from timemachine.fe.dummy import convert_bond_list_to_nx, generate_dummy_group_assignments, get_romol_bonds
-
-    bonds_b = get_romol_bonds(mol_b)
-    bond_graph_b = convert_bond_list_to_nx(bonds_b)
-    dummy_group_assignments = generate_dummy_group_assignments(bond_graph_b, core[:, 1])
-
-    # pick an arbitrary one
-    # later on, iterate over multiple ones.
-    dga = next(dummy_group_assignments)
-
-    for anchor, dg in dga.items():
-        all_idxs, all_params = setup_dummy_bond_and_chiral_interactions_from_ff(
-            ff, mol_b, dg, anchor, core[:, 1], DEFAULT_CHIRAL_ATOM_RESTRAINT_K, DEFAULT_CHIRAL_BOND_RESTRAINT_K
-        )
-        # append idxs
-        all_dummy_bond_idxs.extend(all_idxs[0])
-        all_dummy_chiral_atom_idxs.extend(all_idxs[1])
-        # append params
-        all_dummy_bond_params.extend(all_params[0])
-        all_dummy_chiral_atom_params.extend(all_params[1])
-
-    # generate parameters for mol_a
-    mol_a_top = topology.BaseTopology(mol_a, ff)
-    mol_a_bond_params, mol_a_hb = mol_a_top.parameterize_harmonic_bond(ff.hb_handle.params)
-    mol_a_chiral_atom, mol_a_chiral_bond = mol_a_top.setup_chiral_restraints(
-        DEFAULT_CHIRAL_ATOM_RESTRAINT_K, DEFAULT_CHIRAL_BOND_RESTRAINT_K
-    )
-
-    mol_a_bond_params = mol_a_bond_params.tolist()
-
-    mol_a_bond_idxs = recursive_map(mol_a_hb.idxs, a_to_c)
-    mol_a_chiral_atom_idxs = recursive_map(mol_a_chiral_atom.potential.idxs, a_to_c)
-    mol_a_chiral_bond_idxs = recursive_map(mol_a_chiral_bond.potential.idxs, a_to_c)
-
-    all_dummy_bond_idxs = recursive_map(all_dummy_bond_idxs, b_to_c)
-    all_dummy_chiral_atom_idxs = recursive_map(all_dummy_chiral_atom_idxs, b_to_c)
-
-    # parameterize the combined molecule
-    mol_c_bond_idxs = mol_a_bond_idxs + all_dummy_bond_idxs
-    mol_c_bond_params = mol_a_bond_params + all_dummy_bond_params
-
-    # process chiral volumes, turning off ones at the end-state that have a missing bond.
-    canon_mol_a_bond_idxs_set = set([canonicalize_bond(x) for x in mol_a_bond_idxs])
-    # assert presence of bonds
-    for c, i, j, k in mol_a_chiral_atom_idxs:
-        ci = canonicalize_bond((c, i))
-        cj = canonicalize_bond((c, j))
-        ck = canonicalize_bond((c, k))
-        assert ci in canon_mol_a_bond_idxs_set
-        assert cj in canon_mol_a_bond_idxs_set
-        assert ck in canon_mol_a_bond_idxs_set
-
-    canon_mol_c_bond_idxs_set = set([canonicalize_bond(x) for x in mol_c_bond_idxs])
-
-    # Chiral atom restraint c,i,j,k requires that all bonds ci, cj, ck be present at the
-    # end-state in order to be numerically stable under small perturbations due to normalization
-    # along the bond lengths. However, the angle terms defining icj, ick, and jck can be
-    # either 0 or 180, since the normalized chiral volume is still smooth wrt perturbations
-    all_proper_dummy_chiral_atom_idxs = []
-    all_proper_dummy_chiral_atom_params = []
-    for (c, i, j, k), p in zip(all_dummy_chiral_atom_idxs, all_dummy_chiral_atom_params):
-        ci = canonicalize_bond((c, i))
-        cj = canonicalize_bond((c, j))
-        ck = canonicalize_bond((c, k))
-        if ci in canon_mol_c_bond_idxs_set and cj in canon_mol_c_bond_idxs_set and ck in canon_mol_c_bond_idxs_set:
-            all_proper_dummy_chiral_atom_idxs.append((c, i, j, k))
-            all_proper_dummy_chiral_atom_params.append(p)
-        else:
-            warnings.warn(f"Chiral Volume {c,i,j,k} has a disabled bond, turning off.")
-
-    mol_c_chiral_atom_idxs = list(mol_a_chiral_atom_idxs) + list(all_proper_dummy_chiral_atom_idxs)
-    mol_c_chiral_atom_params = np.concatenate([mol_a_chiral_atom.params, all_proper_dummy_chiral_atom_params])
-
-    # canonicalize bonds
-    mol_c_bond_idxs_canon = np.array([canonicalize_bond(idxs) for idxs in mol_c_bond_idxs])
-    bond_potential = HarmonicBond(mol_c_bond_idxs_canon).bind(np.array(mol_c_bond_params))
-
-    # chiral atoms need special code for canonicalization, since triple product is invariant
-    # under rotational symmetry (but not something like swap symmetry)
-    canon_chiral_atom_idxs = []
-    for idxs in mol_c_chiral_atom_idxs:
-        canon_chiral_atom_idxs.append(canonicalize_chiral_atom_idxs(idxs))
-
-    chiral_atom_idxs = np.array(canon_chiral_atom_idxs, dtype=np.int32).reshape((-1, 4))
-    mol_c_chiral_bond_idxs_canon = [canonicalize_bond(idxs) for idxs in mol_a_chiral_bond_idxs]
-    chiral_bond_idxs = np.array(mol_c_chiral_bond_idxs_canon, dtype=np.int32).reshape((-1, 4))
-    chiral_bond_signs = np.array(mol_a_chiral_bond.potential.signs)
-
-    chiral_atom_potential = ChiralAtomRestraint(chiral_atom_idxs).bind(mol_c_chiral_atom_params)
-    chiral_bond_potential = ChiralBondRestraint(chiral_bond_idxs, chiral_bond_signs).bind(mol_a_chiral_bond.params)
-
-    num_atoms = mol_a.GetNumAtoms() + mol_b.GetNumAtoms() - len(core)
-    assert (
-        get_num_connected_components(num_atoms, bond_potential.potential.idxs) == 1
-    ), "hybrid molecule has multiple connected components"
-    return bond_potential, chiral_atom_potential, chiral_bond_potential
 
 
 def setup_end_state(ff, mol_a, mol_b, core, a_to_c, b_to_c):
@@ -1030,105 +781,6 @@ class AtomMapMixin:
 
 
 _Bonded = TypeVar("_Bonded", bound=Union[ChiralAtomRestraint, HarmonicAngleStable, HarmonicBond, PeriodicTorsion])
-
-
-def get_neighbors(atom, bond_idxs) -> List[int]:
-    nbs = []
-    for i, j in bond_idxs:
-        if i == atom:
-            nbs.append(j)
-        elif j == atom:
-            nbs.append(i)
-    return nbs
-
-
-def check_chiral_validity(src_chiral_centers_in_mol_c, dst_chiral_restr_idx_set, src_bond_idxs):
-    """Raise error unless, for every chiral center, at least 1 chiral volume is defined in both end-states."""
-
-    for c in src_chiral_centers_in_mol_c:
-        nbs = get_neighbors(c, src_bond_idxs)
-        if len(nbs) == 4:
-            i, j, k, l = nbs
-            # (ytz): the ordering of i,j,k,l is random if we're reading directly from the mol graph,
-            # which can be inconsistent with the ordering used in the chiral volume definition.
-            nb_subsets = [(i, j, k), (i, j, l), (i, k, l), (j, k, l)]  # 4-choose-3 subsets
-            flags = [dst_chiral_restr_idx_set.defines((c, ii, jj, kk)) for (ii, jj, kk) in nb_subsets]
-
-            if sum(flags) == 0:
-                raise ChiralConversionError(f"len(nbs) == 4 {c, i, j, k, l}")
-
-        if len(nbs) == 3:
-            i, j, k = nbs
-            flag_0 = dst_chiral_restr_idx_set.defines((c, i, j, k))
-            if not flag_0:
-                raise ChiralConversionError(f"len(nbs) == 3 {c, i, j, k}")
-
-
-def assert_chiral_consistency_and_validity(
-    atom_map: AtomMapMixin,
-    src_chiral_idxs,
-    dst_chiral_idxs,
-    src_bond_idxs: NDArray,
-    dst_bond_idxs: NDArray,
-):
-    """
-    Assert that the given the two end states chiral and bond idxs it would be both consistent and valid.
-
-    consistency: if there are no inversions at the end-states between chiral atoms and bonds are present
-    validity: if we can directly turn on the chiral volumes (after bonds) without staggering angles
-    """
-
-    for c, i, j, k in src_chiral_idxs:
-        assert canonicalize_bond((c, i)) in src_bond_idxs
-        assert canonicalize_bond((c, j)) in src_bond_idxs
-        assert canonicalize_bond((c, k)) in src_bond_idxs
-
-    for c, i, j, k in dst_chiral_idxs:
-        assert canonicalize_bond((c, i)) in dst_bond_idxs
-        assert canonicalize_bond((c, j)) in dst_bond_idxs
-        assert canonicalize_bond((c, k)) in dst_bond_idxs
-
-    src_chiral_restr_idx_set = ChiralRestrIdxSet(src_chiral_idxs)
-    dst_chiral_restr_idx_set = ChiralRestrIdxSet(dst_chiral_idxs)
-
-    # ensure that we don't have any chiral inversions between src and dst end states
-    assert len(src_chiral_restr_idx_set.allowed_set.intersection(dst_chiral_restr_idx_set.disallowed_set)) == 0
-    assert len(dst_chiral_restr_idx_set.allowed_set.intersection(src_chiral_restr_idx_set.disallowed_set)) == 0
-
-    chiral_centers_in_mol_a = chiral_utils.find_chiral_atoms(atom_map.mol_a)
-    chiral_centers_in_mol_b = chiral_utils.find_chiral_atoms(atom_map.mol_b)
-
-    src_chiral_centers_in_mol_c = [atom_map.a_to_c[x] for x in chiral_centers_in_mol_a]
-    dst_chiral_centers_in_mol_c = [atom_map.b_to_c[x] for x in chiral_centers_in_mol_b]
-
-    check_chiral_validity(src_chiral_centers_in_mol_c, dst_chiral_restr_idx_set, src_bond_idxs)
-    check_chiral_validity(dst_chiral_centers_in_mol_c, src_chiral_restr_idx_set, dst_bond_idxs)
-
-
-def verify_chiral_consistency_of_core(mol_a: Chem.Mol, mol_b: Chem.Mol, core: NDArray, forcefield):
-    """Verify that a core and forcefield would allow for valid chiral endstates.
-
-    Refer to `assert_chiral_consistency_and_validity` for definitions of consistency and validity.
-
-    Raises
-    ------
-        ChiralConversionError
-            If chiral end states are incompatible for the given core and forcefield.
-    """
-    atom_map = AtomMapMixin(mol_a, mol_b, core)
-    bond_pot_src, chiral_atom_pot_src, _ = setup_end_state_harmonic_bond_and_chiral_potentials(
-        forcefield, mol_a, mol_b, core, atom_map.a_to_c, atom_map.b_to_c
-    )
-    bond_pot_dest, chiral_atom_pot_dest, _ = setup_end_state_harmonic_bond_and_chiral_potentials(
-        forcefield, mol_b, mol_a, core[:, ::-1], atom_map.b_to_c, atom_map.a_to_c
-    )
-    assert_chiral_consistency_and_validity(
-        atom_map,
-        chiral_atom_pot_src.potential.idxs,
-        chiral_atom_pot_dest.potential.idxs,
-        bond_pot_src.potential.idxs,
-        bond_pot_dest.potential.idxs,
-    )
 
 
 class SingleTopology(AtomMapMixin):
